Add files via upload
This is a password entropy program I wrote to measure the n amount of combinations needed to crack a password depending on the length of the chosen password. A typical strong password is considered to be  8 characters in length, so I decided to set my variable of 'good_password_length' to the total number of combinations for an 8 bit password. 
The interesting thing that I took out of this project was that selecting an 8 bit character password such as 12345678 and 84@as912 both give you the same number of combinations needed to crack a password. 

A lot of hacks happen because users don't pick strong passwords and use the same password for many different accounts. This gives an attacker a range of different act avenues to explore when targeting specific targets. Systems are filled with externalizes that affect other systems in unforeseen and potentially harmful ways. Vulnerabilities on one system can cascade into another. 

If you would like to know more about online security, I would recommenced visiting the website for the Australian Cyber Security Center for a range of advice on how to secure your systems and how navigate the internet safely. 
I hope you enjoyed my project and that you learnt a thing or two. 

https://www.cyber.gov.au/acsc/view-all-content/advice